Initial Release
Three weeks of coding has brought me to this point. A lot of this was aided by ChatGPT, no I don't care, I used it as a tool to learn, I didn't just have it write the code for me. This is the first release of the WurliMac, a vintage-inspired jukebox project that turns your classic Mac, any PPC machine capable of running Tiger, into a jukebox that actually takes requests like a real diner jukebox!