Add files via upload
The first "real" version of the WurliMac. It has a proper(ish) class definition, a much cleaner loop, and more configuration options. Still no stop or shutdown functions, but I'll add those sooner or later, I promise! I would encourage anyone to put this on their Tiger Mac and give it a try. Remember to get an updated version of Python!